ui/ops/accounts: move role/account functionality closer to those pages
