+++ ui/ops/src/routes/auth/roles/roles.js
@@ -1,5 +1,48 @@
+import {getRole, listRoles} from '@/api/ui/account.js';
+import {useAction} from '@/composables/action.js';
+import useCollection from '@/composables/collection.js';
 import {usePermissionsStore} from '@/stores/permissions.js';
 import {computed, toValue} from 'vue';
+
+/**
+ * @param {import('vue').MaybeRefOrGetter<Partial<ListRolesRequest.AsObject>>} request
+ * @param {import('vue').MaybeRefOrGetter<Partial<ListOnlyCollectionOptions<ListRolesRequest.AsObject, Role.AsObject>>>?} options
+ * @return {UseCollectionResponse<Role.AsObject>}
+ */
+export function useRolesCollection(request, options) {
+  const normOpts = computed(() => {
+    return {
+      cmp: (a, b) => a.id.localeCompare(b.id, undefined, {numeric: true}),
+      ...toValue(options)
+    };
+  });
+  const client = {
+    async listFn(req, tracker) {
+      const res = await listRoles(req, tracker);
+      return {
+        items: res.rolesList,
+        nextPageToken: res.nextPageToken,
+        totalSize: res.totalSize
+      };
+    },
+    pullFn(req, resource) {
+      const opts = toValue(normOpts);
+      if (opts.pullFn) {
+        opts.pullFn(req, resource);
+      }
+    }
+  };
+
+  return useCollection(request, client, normOpts);
+}
+
+/**
+ * @param {import('vue').MaybeRefOrGetter<Partial<GetRoleRequest.AsObject>>} request
+ * @return {ToRefs<UnwrapNestedRefs<UseActionResponse<Role.AsObject>>>}
+ */
+export function useGetRole(request) {
+  return useAction(request, getRole);
+}
 
 /**
  * @typedef {Permission.AsObject} AssignedPermission
